backend/requirements.txt
flask
pymongo
flask-cors
python-dotenv
werkzeug
google-genai
cloudinary
bcrypt
